add sound gen file and add support for multiple sound sources
